# flowdesk_ready/17_flowdesk_api_contract_v0.8.yaml
openapi: 3.1.0
info:
  title: FlowDesk Live Transaction API
  version: 0.8.0
  description: FlowDesk Live transaction contract through C12. PostgreSQL remains
    the durable transaction authority. C09 enforces stale-stock preconditions, C11
    provides atomic reservation variant switching, and C12 introduces a deterministic
    closing state that stops new buyer intents, cancels waiting buyers, drains existing
    holds by their original server-authoritative TTLs, suppresses further waitlist
    promotions, and ends only with zero active/payment-pending reservations and zero
    waiting entries.
security:
- bearerAuth: []
...
paths:
...
  /v1/reservations/{reservationId}/cancel:
    post:
      tags:
      - Reservations
      summary: Cancel an active reservation and release stock exactly once
      operationId: cancelReservation
      parameters:
      - $ref: '#/components/parameters/ReservationId'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Reservation cancellation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /v1/operator-overrides/force-allocation:
    post:
      tags:
      - Overrides
      summary: Privileged audited force-allocation with commit-time stock recheck
      operationId: forceAllocationOverride
      x-flowdesk-permission: reservation.force_allocate
      x-flowdesk-step-up-max-age-seconds: 600
      description: A force allocation bypasses normal queue policy only; it never
        bypasses authoritative stock. The request must carry If-Inventory-Version
        from a fresh client snapshot. The server locks the inventory row, compares
        that version, rechecks available quantity, and only then may create the reservation.
        Successful mutation and audit event commit atomically.
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/InventoryVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForceAllocationRequest'
      responses:
        '201':
          description: Override committed and audited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '403':
          description: Actor lacks privileged role or step-up authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          $ref: '#/components/responses/Conflict'
        '428':
          description: If-Inventory-Version is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/reservations/{reservationId}/switch-variant:
    post:
      tags:
      - Reservations
      summary: Atomically move an active reservation to another inventory variant
      operationId: switchReservationVariant
      description: Locks the reservation first, then locks source and destination
        inventory rows in deterministic ID order. Source inventory is released only
        if destination inventory can be claimed in the same transaction. On failure,
        the original reservation remains unchanged. A successful switch writes an
        immutable audit event and queues source-stock waitlist work before COMMIT.
      parameters:
      - $ref: '#/components/parameters/ReservationId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VariantSwitchRequest'
      responses:
        '200':
          description: Variant switch committed or idempotent no-change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariantSwitchResult'
        '409':
          $ref: '#/components/responses/Conflict'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 200
      description: Merchant-scoped mutation key. Same key + same payload returns the
        original result; same key + different payload is a conflict.
...
    ReservationId:
      name: reservationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    InventoryVersion:
      name: If-Inventory-Version
      in: header
      required: true
      schema:
        type: integer
        format: int64
        minimum: 0
      description: Version from the seller client's fresh, session/SKU-scoped authoritative
        stock snapshot. The server locks the inventory row and rejects a mismatch
        before any reservation or audit mutation.
  responses:
    Conflict:
      description: Current authoritative state prevents the requested mutation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      - request_id
      properties:
        code:
          type: string
          enum:
          - INVENTORY_UNAVAILABLE
          - STALE_INVENTORY_VERSION
          - IDEMPOTENCY_KEY_REUSED_WITH_DIFFERENT_PAYLOAD
          - RESERVATION_ALREADY_TERMINAL
          - WAITLIST_ENTRY_NOT_ELIGIBLE
          - SESSION_NOT_LIVE
          - PREFLIGHT_FAILED
          - RECONCILIATION_REQUIRED
          - FORBIDDEN_OVERRIDE
          - INVENTORY_VERSION_REQUIRED
          - INVALID_INVENTORY_VERSION
          - DESTINATION_VARIANT_UNAVAILABLE
          - DESTINATION_OWNERSHIP_CONFLICT
          - SOURCE_INVENTORY_NOT_FOUND
          - DESTINATION_INVENTORY_NOT_FOUND
        message:
          type: string
        request_id:
          type: string
        current_inventory_version:
          type: integer
          format: int64
...
    Reservation:
      type: object
      additionalProperties: false
      required:
      - id
      - live_session_id
      - inventory_variant_id
      - buyer_id
      - quantity
      - status
      - expires_at
      - created_at
      - updated_at
      properties:
        id:
          type: string
          format: uuid
        live_session_id:
          type: string
          format: uuid
        inventory_variant_id:
          type: string
          format: uuid
        buyer_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        status:
          type: string
          enum:
          - active
          - payment_pending
          - paid
          - expired
          - cancelled
          - reconciliation_required
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ForceAllocationRequest:
      type: object
      additionalProperties: false
      required:
      - live_session_id
      - inventory_variant_id
      - buyer_id
      - quantity
      - reason
      - step_up_token
      properties:
        live_session_id:
          type: string
          format: uuid
        inventory_variant_id:
          type: string
          format: uuid
        buyer_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        reason:
          type: string
          minLength: 8
          maxLength: 1000
        step_up_token:
          type: string
          minLength: 16
...
    VariantSwitchRequest:
      type: object
      additionalProperties: false
      required:
      - destination_inventory_variant_id
      - reason
      properties:
        destination_inventory_variant_id:
          type: string
          format: uuid
        reason:
          type: string
          minLength: 8
          maxLength: 500
          description: Human-readable buyer/operator intent captured in immutable
            audit history.
    VariantSwitchResult:
      type: object
      additionalProperties: false
      required:
      - outcome
      - reservation_id
      - quantity
      properties:
        outcome:
          type: string
          enum:
          - switched
          - no_change
        reservation_id:
          type: string
          format: uuid
        from_inventory_variant_id:
          type:
          - string
          - 'null'
          format: uuid
        to_inventory_variant_id:
          type:
          - string
          - 'null'
          format: uuid
        quantity:
          type: integer
          minimum: 1
        source_inventory_version:
          type:
          - integer
          - 'null'
          format: int64
        destination_inventory_version:
          type:
          - integer
          - 'null'
          format: int64
        promotion_queued:
          type:
          - boolean
          - 'null'
...